Update to Scala 2.12.18 (#2119)
* Update from Scala 2.12.17 to 2.12.18. This includes a update from ASM 9.3 to 9.5, which improved JDK 20 and 21 support.
* Fix integral division widening to float warnings.  In this commit, the new lint warning introduced with Scala 2.12.18, which pertains to integer divisions that are implicitly converted to floats, is addressed.

+++ netlogo-core/src/main/shape/Circle.scala
@@ -37,7 +37,7 @@ class Circle(color: Color) extends Element(color) with BaseCircle with Cloneable
   }
 
   def origin: Point =
-    new Point(x + round(xDiameter / 2), y + round(yDiameter / 2))
+    new Point(x + round(xDiameter.toFloat / 2), y + round(yDiameter.toFloat / 2))
 
   def bounds: AwtRectangle =
     new AwtRectangle(x, y, xDiameter, yDiameter)
