+++ netlogo-gui/src/main/prim/hubnet/primitives.scala
@@ -192,7 +192,7 @@ class _hubnetresetperspective extends Command with HubNetPrim {
         override def run() {
           hubNetManager.foreach(_.sendAgentPerspective(
             client, world.observer.perspective.export,
-            agentKind, id, (world.worldWidth - 1) / 2, true))
+            agentKind, id, ((world.worldWidth - 1).toDouble / 2), true))
         }})
     context.ip = next
   }
